Adventures in recursion.
The definition of the parser for the ab language is a bit more verbose
than one would like it to be. However, it works.

The same is not true for the following, simpler definition (which the
type checker accepts):

    ab = unit | ('a'_P & ab & 'b'_P & ab) | ('b'_P & ab & 'a'_P & ab);

Running this will result in `std::bad_function_call`. What's wrong?