Added eta expansion macro.
The problem with the naive definition

    Parser ab = unit | 'a'_P & ab & 'b'_P & ab | 'b'_P & ab & 'a'_P & ab;

of the grammar for the ab language was essentially due to the fact that
in this definition the lhs---although it is uninitialized yet---is still
available to the rhs. Thus, we use an uninitialized value of type Parser
to define the lhs, which obviously results in nonsense.

The easiest (although somewhat dangerous-looking) solution is to
eta-expand the rhs.

For that purpose I added a macro `ETA(x)`. Perhaps I should also add
a fix point combinator at some point in order to write anonymous
recursive Parser objects (or to be a bit more conceptual).